Move contract events into events folder
